news 타입 정리
