+++ apis/news.ts
@@ -1,21 +1,15 @@
-import {
-  PostSearchQueryParams,
-  GETNewsPostsResponse,
-  NewsPostResponse,
-  NewsPost,
-  POSTNewsBody,
-  PATCHNewsBody,
-} from '@/types/post';
+import { News, NewsPreviewList, PATCHNewsBody, POSTNewsBody } from '@/types/news';
+import { PostSearchQueryParams } from '@/types/post';
 
 import { deleteRequest, getRequest, patchRequest, postRequest } from '.';
 
 const newsPath = '/news';
 
 export const getNewsPosts = (params: PostSearchQueryParams) =>
-  getRequest(newsPath, params, { cache: 'no-store' }) as Promise<GETNewsPostsResponse>;
+  getRequest(newsPath, params, { cache: 'no-store' }) as Promise<NewsPreviewList>;
 
 export const getNewsPostDetail = (id: number, params?: PostSearchQueryParams) =>
-  getRequest(`${newsPath}/${id}`, params, { cache: 'no-store' }) as Promise<NewsPostResponse>;
+  getRequest(`${newsPath}/${id}`, params, { cache: 'no-store' }) as Promise<News>;
 
 export const postNews = async (body: POSTNewsBody) => {
   const formData = new FormData();
@@ -35,9 +29,9 @@ export const postNews = async (body: POSTNewsBody) => {
     formData.append('attachments', attachment);
   }
 
-  return (await postRequest(newsPath, {
+  await postRequest(newsPath, {
     body: formData,
-  })) as NewsPostResponse;
+  });
 };
 
 export const patchNews = async (id: number, body: PATCHNewsBody) => {
