+++ components/news/NewsPageContent.tsx
@@ -10,8 +10,8 @@ import { NewsTags } from '@/constants/tag';
 import { useCustomSearchParams } from '@/hooks/useCustomSearchParams';
 import { useQueryString } from '@/hooks/useQueryString';
 
+import { NewsPreviewList } from '@/types/news';
 import { news } from '@/types/page';
-import { GETNewsPostsResponse } from '@/types/post';
 
 import { getPath } from '@/utils/page';
 
@@ -21,7 +21,7 @@ const newsPath = getPath(news);
 export default function NewsPageContent({
   data: { searchList, total },
 }: {
-  data: GETNewsPostsResponse;
+  data: NewsPreviewList;
 }) {
   const { page, keyword, tags, setSearchParams } = useCustomSearchParams();
   const queryString = useQueryString();
